+++ lib/src/step/completion_step.dart
@@ -82,9 +82,7 @@ class CompletionStep extends FormStep {
   ) {
     final json = JsonReader(element, context: 'CompletionStep');
     return CompletionStep(
-      display: json.string('display') != null
-          ? Display.values.firstWhere((e) => e.name == json.string('display'))
-          : Display.normal,
+      display: json.enumValue('display', Display.values) ?? Display.normal,
       crossAxisAlignmentContent:
           crossAlignmentFromString(
             json.string('crossAxisAlignmentContent') ?? "center",
@@ -94,10 +92,10 @@ class CompletionStep extends FormStep {
       autoTrigger: json.boolean('autoTrigger') ?? false,
       style: UIStyle.maybeFrom(json.map('style')),
       relevantConditions: relevantConditions,
-      backButtonText: json.string('backButtonText'),
-      cancelButtonText: json.string('cancelButtonText'),
+      backButtonText: json.string('backButtonText') ?? "Back",
+      cancelButtonText: json.string('cancelButtonText') ?? "Cancel",
       isOptional: json.boolean('isOptional'),
-      nextButtonText: json.string('nextButtonText'),
+      nextButtonText: json.string('nextButtonText') ?? "Finish",
       successLottieAssetsFilePath: json.string('successLottieAssetsFilePath'),
       loadingLottieAssetsFilePath: json.string('loadingLottieAssetsFilePath'),
       errorLottieAssetsFilePath: json.string('errorLottieAssetsFilePath'),