+++ lib/src/step/instruction_step.dart
@@ -52,9 +52,7 @@ class InstructionStep extends FormStep {
   ) {
     final json = JsonReader(element, context: 'InstructionStep');
     return InstructionStep(
-      display: json.string('display') != null
-          ? Display.values.firstWhere((e) => e.name == json.string('display'))
-          : Display.normal,
+      display: json.enumValue('display', Display.values) ?? Display.normal,
       crossAxisAlignmentContent:
           crossAlignmentFromString(
             json.string('crossAxisAlignmentContent') ?? "center",
@@ -63,13 +61,13 @@ class InstructionStep extends FormStep {
       style: UIStyle.maybeFrom(json.map('style')),
       cancellable: json.boolean('cancellable'),
       relevantConditions: relevantConditions,
-      backButtonText: json.string('backButtonText'),
-      cancelButtonText: json.string('cancelButtonText'),
+      backButtonText: json.string('backButtonText') ?? "Back",
+      cancelButtonText: json.string('cancelButtonText') ?? "Cancel",
       isOptional: json.boolean('isOptional'),
       instructions: DynamicData.parseDynamicData(
         cast<List<dynamic>>(element?["instructions"]) ?? const [],
       ),
-      nextButtonText: json.string('nextButtonText'),
+      nextButtonText: json.string('nextButtonText') ?? "Start",
       text: json.string('text'),
       title: json.string('title'),
       titleIconAnimationFile: json.string('titleIconAnimationFile'),